feat: Add ui-dialog component (#42)
* feat: Add ui-modal component

This commit adds the ui-modal component to the @halvaradop/ui library. The ui-modal component provides a modal package for the library. It includes a new package.json file with scripts for building and cleaning, as well as a tsconfig.json file and a tsup.config.ts file. Additionally, it introduces a new index.tsx file that exports the Modal component and a modal.stories.tsx file with different stories showcasing the component's usage.

* refactor: clean up stories and rename package

- Rename package from `ui-modal` to `ui-dialog`
- Create a custom component to eliminate duplicate code in stories

* chore: install dependencies in `ui-dialog`